Clean up on ProposalDetailsScreen
