+++ lib/v2/screens/explore_screens/vote_screens/proposal_details/viewmodels/proposal_details_bloc.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/screens/explore_screens/vote_screens/proposal_details/viewmodels/bloc.dart';
 import 'package:seeds/v2/screens/explore_screens/vote_screens/proposal_details/viewmodels/page_commands.dart';
 import 'package:seeds/v2/screens/explore_screens/vote_screens/proposals/viewmodels/proposals_and_index.dart';
@@ -9,6 +10,11 @@ class ProposalDetailsBloc extends Bloc<ProposalDetailsEvent, ProposalDetailsStat
 
   @override
   Stream<ProposalDetailsState> mapEventToState(ProposalDetailsEvent event) async* {
+    if (event is OnLoadProposalData) {
+      yield state.copyWith(pageState: PageState.loading);
+      // next pr load logic
+      yield state.copyWith(pageState: PageState.success);
+    }
     if (event is OnNextProposalTapped) {
       yield state.copyWith(
         pageCommand: ReturnToTopScreen(),
